+++ src/types.h
@@ -2,6 +2,6 @@
 #define TYPES_H_
 
 #include <boost/numeric/ublas/vector.hpp>
-typedef boost::numeric::ublas::vector<float> Weights;
+typedef boost::numeric::ublas::vector<double> Weights;
 
 #endif  // TYPES_H_
